requirements.txt
Django==1.6.5
django-haystack==2.2.0
elasticsearch==1.2.0
pyelasticsearch==0.6.1
requests==2.3.0
simplejson==3.6.0
six==1.7.3
urllib3==1.9

# you may also need to have ELASTICSEARCH running on the machin for it to work